ml/requirements.txt
numpy
Pillow
matplotlib
tensorflow
